 requirements.txt
flask
pandas
mysql-connector-python
scikit-learn
numpy
matplotlib
warnings
requests